+++ future/tests/test_futurize.py
@@ -32,6 +32,7 @@ class TestFuturizeSimple(CodeHandler):
         """
         self.convert_check(before, after)
 
+    @unittest.skip('--tobytes feature removed for now ...')
     def test_tobytes(self):
         """
         The --tobytes option converts all UNADORNED string literals 'abcd' to b'abcd'.
@@ -334,7 +335,6 @@ class TestFuturizeRenamedStdlib(CodeHandler):
         """
         self.convert_check(before, after)
     
-    @unittest.expectedFailure
     def test_urllib_refactor(self):
         # Code like this using urllib is refactored by futurize --stage2 to use
         # the new Py3 module names, but ``future`` doesn't support urllib yet.
@@ -347,7 +347,7 @@ class TestFuturizeRenamedStdlib(CodeHandler):
         data = r.read()
         """
         after = """
-        import urllib.request
+        import urllib.request, urllib.parse, urllib.error
         
         URL = 'http://pypi.python.org/pypi/future/json'
         package_name = 'future'