+++ libfuturize/test_scripts/py3/implrelimport_printfunction.py
@@ -1,3 +1,0 @@
-import emptymodule
-print('Hello', ' Ed', file=sys.stderr)
-